Suggested revised edition to review
Removed the != sunday checks, as all that kind of stuff is handled by the heap structure. Replaced !=00:00 check with start == close check, and added it for all days rather than just sunday. Additionally you could consider using the .split("|") method (for example) on the string response to be more secure in handling opening hours, but that one is merely a suggestion. Also fixed the indentation in the sample response, and renamed titles. Hope this feedback helps :)